Added useKeyboard hook to consolidate duplicated code
