+++ src/js/components/Tip/Tip.js
@@ -4,30 +4,18 @@ import React, {
   forwardRef,
   useContext,
   useState,
-  useEffect,
 } from 'react';
 import { ThemeContext } from 'styled-components';
 
 import { Box } from '../Box';
 import { Drop } from '../Drop';
-import { useForwardedRef } from '../../utils/refs';
+import { useForwardedRef, useKeyboard } from '../../utils';
 import { TipPropTypes } from './propTypes';
 
 const Tip = forwardRef(({ children, content, dropProps, plain }, tipRef) => {
   const theme = useContext(ThemeContext);
   const [over, setOver] = useState(false);
-  const [usingKeyboard, setUsingKeyboard] = useState();
-
-  const onMouseDown = () => setUsingKeyboard(false);
-  const onKeyDown = () => setUsingKeyboard(true);
-  useEffect(() => {
-    document.addEventListener('mousedown', onMouseDown);
-    document.addEventListener('keydown', onKeyDown);
-    return () => {
-      document.removeEventListener('mousedown', onMouseDown);
-      document.removeEventListener('keydown', onKeyDown);
-    };
-  }, []);
+  const usingKeyboard = useKeyboard();
 
   const componentRef = useForwardedRef(tipRef);
