+++ src/js/utils/index.js
@@ -11,3 +11,4 @@ export * from './pagination';
 export * from './PortalContext';
 export * from './refs';
 export * from './responsive';
+export * from './use-keyboard';
